Removed temp generated fsharpi file
First, it seems that this is not necessary, probably because FAKE 3.0 uses the
F# compiler services instead of the fsharpi executable.

Second, because the FAKE command was not the last command in the script, any
failure from the build script would not cause the 'build.sh' shell script to
return an error message, and e.g. a build server would see the build outcome as
a success.