Fix proxy script use
When using a proxy configuration script (aka wpad pac script), the actual
proxy settings you get depend on the url you target.
It's used for load balancing / web proxy routing in enterprises.

If you were unlikely enough (like me) to have a split between google urls
and the rest of the www, the current code would use the google proxy
to target github urls, which was failing (405 in my case but I guess
that could depend on the actual setup).