Add parsing and handling of MSBuild conditions
Currently does NOT handle the `Exists` or `HasTrailingSlash` features,
nor grouping of conditions with parens. The MSDN article on MSBuild
conditions also makes mention of hex literals, which have not been
handled here either. See the following page for more information on these
features:
https://msdn.microsoft.com/en-us/library/7szfhaft.aspx

Line 737 can be changed to `doComp left right` to support performing
greater-than and lesser-than style comparisons on strings, but this is
not supported by MSBuild.