Update ProjectFile.GetOutputDirectory to use given build configuration
The provided configuration needs to exist in the default placeholder
data, so that if it's provided, the value isn't overridden. If the value
is empty, project files typically contain something like this to provide
a default value:

<Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>

With this change, all tests appear to pass again, should be the last of the
changes needed for #914.