Use GetProperty from GetTargetProfile when a profile isn't explictly set
If a <TargetFrameworkProfile> value can be extracted from the project
file and it has a value, that value is used to determine the profile
here. If that node isn't specified, or its value is empty,
GetTargetProfile now tries to get the <TargetFrameworkVersion> node to
determine the correct profile from there.