Add `--force-nuget` CLA to bootstrapper w/`ForceNuget` setting option
If the `--force-nuget` command-line argument is supplied, or the
`ForceNuget` setting exists in the application settings and is set to
`true`, then the download strategy is set to only use the specified
Nuget source. As the previous commit mentioned, if this CLA isn't
specified and the app setting is either not present or not equal to
`true`, the `--prefer-nuget` option is evaluated as it was before.