ProjectFile.touch to only modify the last write time without changing file content
This is a follow up to #1485 which introduced the optional --touch-affected-refs
parameter on restore. In the original implementation this caused the
project file to be saved forcefully, which could in some cases cause the
file content to change slightly (e.g. minor xml formatting).

With this change, the only effect of --touch-affected-refs will be a
modified last-write time file attribute.