Added convert-from-nuget environment vars note
When converting a project from NuGet to Paket, it is possible to get into a state that a paket.dependencies file is generated, but the process is not finished. If you put plaintext credentials in this spot, it will continue as suspected with the --hard argument. If you put an environment variable it is not expanded.

As this situation might be confusing to anyone who gets into this situation and sees this document, a note linking to the proper credential migration procedure is a reasonable concession.

This references issue 1822.