Add a "magic" mode to the bootstraper
If the bootstraper is renamed to paket.exe it download the real paket.exe
to the temporary folder instead.

In addition if --run isn't specified on the command line it act as if it
was passed and run paket.exe with all the arguments it received.

The result is that it produce a paket.exe of 47Ko that act like the normal
one (As long as an internet connection is available)