Add Install-Paket.ps1 to prevent embedding .exe in git.
This checks the github api for the latest release and appropriately
downloads the paket bootstrapper. An optional parameter -AddToPath can
be used to add the paket directory to $env:Path for CI boxes.

The intent is to allow those who have *.exe in their .gitignore to more
easily use paket on Windows.