Add a concept of 'Verbosity' in the boostrapper
It replace the 'Silent' that was present before. Verbosity can be lowered
by one or more '-s' and raised by one or more '-v'.

This commit also change the behavior of the boostrapper when called with a
single '-s' argument, it's now displaying errors only instead of being
totally silent (A second '-s' is now necessary for that)