NuGetV2-OData: retrieve versions in descending order for artifactory
Related to #2008 which was reverted in #2018 because it affected non-artifactory
servers negatively (in an unknonwn way).

This change adds an alternative variant of tryGetAllVersions to fetch the
versions in descending order (by publish date) instead of modifying the
original one.

GetVersions chooses to use the alternative variant in case the source uri
contains the string "artifactory". This is a very unreliable way of
discovering a server to be an artifactory one, but is in line with the
existing rules and should prevent affecting non-artifactory servers.

Ideally the decision should be more controllable than url matching, e.g.
by introducing the concept of source options.