new option generate_load_scripts: on|off|true|false in paket.dependencies
this option, if on/true, calls the loading script generation

Note that this does it in very eager way (if defined for any group, it will generate all), it also doesn't allow to specify framework or language like from command line, keeping it simple for now.

The option is off by default, because there is performance penality and it generates lots of small script files, there is also no proper logic for "garbage collection" process.

Implementation notes:
* A new convenience method is also exposed on PublicAPI: Dependencies.GenerateIncludeScripts()
* I had to migrate a good deal of stuff in a module I named Queries, mostly stuff which were raw in Program.fs
* added a unit test for parser, testing all cases of the new option
* added an integration checking calling install with option on results in script being generated without further invocation
* Since both .dependencies and .lock files are used often in conjunction, I made a PaketFiles DU (in Queries module)
* Since both group name and package name are used often in conjunction, I made a QualifiedPackageName single case DU, which helped me remove a bit of cruft related to default group