`paket/lock` tag is created and checked out in the downstream repository when `update`ing/`restore`ing the dependency
This allows user to see what the original checked out commit was when making changes.
By checking out the tag - instead of resetting the downstream repo `--hard` to the tracking branch - the repo starts in the detached HEAD state which:
- allows easy experimentation - user can discard her commits simply by checking out the `paket/lock` tag again
- forces user to checkout a branch when doing meaningful changes.
Note that the tracking branch for each downstream repo (`b<REPO_PATH_HASH>`) is no longer created.