Downstream repository is configured with the `origin` remote pointing to the upstream repository (instead of the cache repository)
Having the upstream repo configured as the `origin` makes pushing changes to the upstream more natural.
We also want to prevent user from pushing the changes to the cache directly - the cache should be managed by Paket.
Note that fetching from the cache has been updated to use explicit refspec. This way we do not need an entry with matching URL in gitconfig.