Add support for dotnet SDK fallback folder
The dotnet SDK ships with various .NET Core and ASP.NET Core nuget
packages in a NuGetFallbackFolder.

This changes to try to find packages in that folder before looking in
other caches or downloading from source. In the event the package is
found here, the contents are copied or linked directly from that folder
instead of the user's cache.