paket.targets -> Paket.Restore.targets
I believe the docs are out of date on this file. I don't have a paket.targets file, only a Packet.Restore.targets file. Updated the docs on the assumption Paket.Restore.targets is the new name.

Pls review my comment on the implications of not committing this file as I'm not sure that is correct.