Update git-dependencies.md
with #3756 and #3064 it is quite clear, that the documentation around adding git repositories as a (usable) dependency isn't all that clear. Which is a pity, because the actual process isn't all that hard and also quite sensible-ish.

This change aims to slightly improve the docs and shed some light on this topic because - basically - all that is missing is some basic knowledge about how namespace/package resolution works in fsharp.